feat: polish FrameValidationData
